requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
flask==3.0.0
flask-cors==4.0.0
scipy==1.11.4
Pillow==10.1.0
lxml==4.9.3
opencv-python==4.8.1.78 